Added rank ID to the problem struct
- There is a bug in the current ISAT implementation caused by interacting a problem with the ISAT tree of its host rank.
Therefore only interaction with the ISAT is now performed for problems that were not balanced (i.e., stays in its original rank).
This is ensured by adding a check that compares the rank ID of the problem with the current rank ID, and goes through tabulation
only if that is satisfied.